fix: use config
