+++ src/config/index.ts
@@ -1,0 +1,10 @@
+export const HOME_NAVIGATION = [
+  { name: "Sign in", href: "/login" },
+  { name: "Github", href: "https://github.com/Bartek532/rssmarkable" },
+];
+
+export const HOME_FOOTER_NAVIGATION = [
+  { name: "Github", href: "https://github.com/Bartek532/rssmarkable" },
+  { name: "Login/Signup", href: "/login" },
+  { name: "Dashboard", href: "/dashboard" },
+];
